feedback couselor

+++ app/src/main/java/com/tdc/edu/vn/heathcareapp/Model/Counselor.java
@@ -2,13 +2,13 @@ package com.tdc.edu.vn.heathcareapp.Model;
 
 public class Counselor {
     String id_counselor, user_id, category, introduce, position_counselor;
-    int total_ratting;
+    Float total_ratting;
 
     public Counselor() {
 
     }
 
-    public Counselor(String id_counselor, String user_id, String category, String introduce, String position_counselor, int total_ratting) {
+    public Counselor(String id_counselor, String user_id, String category, String introduce, String position_counselor, Float total_ratting) {
         this.id_counselor = id_counselor;
         this.user_id = user_id;
         this.category = category;
@@ -57,11 +57,11 @@ public class Counselor {
         this.position_counselor = position_counselor;
     }
 
-    public int getTotal_ratting() {
+    public Float getTotal_ratting() {
         return total_ratting;
     }
 
-    public void setTotal_ratting(int total_ratting) {
+    public void setTotal_ratting(Float total_ratting) {
         this.total_ratting = total_ratting;
     }
 }
